requirements.txt
# --- Ingestion ---
langchain>=0.2.0
langchain-community>=0.2.0
atlassian-python-api>=3.41.0        # Confluence loader dependency
Office365-REST-Python-Client>=2.5.0 # SharePoint loader dependency

# --- Chunking ---
tiktoken>=0.7.0

# --- Embeddings ---
openai>=1.30.0                      # Azure OpenAI or OpenAI-compatible endpoint
boto3>=1.34.0                       # AWS Bedrock (optional)
sentence-transformers>=3.0.0        # Self-hosted BGE fallback

# --- Vector store ---
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sqlalchemy>=2.0.0

# --- Retrieval ---
rank-bm25>=0.2.2

# --- Reranking ---
# sentence-transformers already listed above (ms-marco cross-encoder)

# --- Generation ---
anthropic>=0.40.0
# openai already listed above

# --- Evaluation ---
ragas>=0.1.0
datasets>=2.20.0

# --- API ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# --- Frontend ---
streamlit>=1.35.0

# --- Teams bot ---
botbuilder-core>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.9.0

# --- Utilities ---
python-dotenv>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.0                 # Nightly re-index scheduler
structlog>=24.1.0                   # Structured audit logging
